updated Pages and app
