+++ src/index.css
@@ -3,23 +3,10 @@ html, body {
   padding: 0;
   background-color: rgb(242, 246, 248);
 }
-
-.App{
-  height: 100vh;
-  display: flex;
-  flex-direction: column;
+*{
+  box-sizing: border-box;
 }
 
 #root{
   text-align: center;
-}
-
-main {
-  display: flex;
-  width: 100%;
-  height: 100%;
-}
-
-.task-section{
-  width: 100%;
 }
