+++ src/main.jsx
@@ -1,10 +1,21 @@
+// Libraries
 import React from 'react'
 import ReactDOM from 'react-dom/client'
+import { QueryClientProvider, QueryClient } from 'react-query'
+
+// App component
 import App from './App'
+
+// Styling
 import './index.css'
 
+const queryClient = new QueryClient();
+
+
 ReactDOM.createRoot(document.getElementById('root')).render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>
+  <QueryClientProvider client = {queryClient}>
+    {/* <React.StrictMode> */}
+      <App />
+    {/* </React.StrictMode> */}
+  </QueryClientProvider>
 )
