+++ src/pages/LoginPage/Login.css
@@ -47,6 +47,16 @@
   padding-bottom: 80px;
 }
 
+.login .login-response-message{
+  background-color: rgb(214, 178, 1);
+  padding: 10px 20px 10px 20px;
+  border-radius: 20px;
+  color: white;
+  margin-bottom: 10px;
+  width: 100%;
+  box-sizing: border-box;
+}
+
 .login .input-field{
   display: flex;
   flex-direction: column;
@@ -82,6 +92,9 @@
 	background-image: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='20' height='20' fill='currentColor' class='bi bi-key' viewBox='0 0 16 16'%3E%3Cpath d='M0 8a4 4 0 0 1 7.465-2H14a.5.5 0 0 1 .354.146l1.5 1.5a.5.5 0 0 1 0 .708l-1.5 1.5a.5.5 0 0 1-.708 0L13 9.207l-.646.647a.5.5 0 0 1-.708 0L11 9.207l-.646.647a.5.5 0 0 1-.708 0L9 9.207l-.646.647A.5.5 0 0 1 8 10h-.535A4 4 0 0 1 0 8zm4-3a3 3 0 1 0 2.712 4.285A.5.5 0 0 1 7.163 9h.63l.853-.854a.5.5 0 0 1 .708 0l.646.647.646-.647a.5.5 0 0 1 .708 0l.646.647.646-.647a.5.5 0 0 1 .708 0l.646.647.793-.793-1-1h-6.63a.5.5 0 0 1-.451-.285A3 3 0 0 0 4 5z'/%3E%3Cpath d='M4 8a1 1 0 1 1-2 0 1 1 0 0 1 2 0z'/%3E%3C/svg%3E");
   background-repeat: no-repeat;
   background-position: 8px 9px;
+}
+.login #error {
+  border: 1px solid red;
 }
 
 .login .password-input{
