+++ src/pages/LoginPage/Login.jsx
@@ -1,43 +1,67 @@
+// Libraries
+import { useQuery } from 'react-query'
 import React,{ useState } from 'react';
 import { Link, useNavigate } from 'react-router-dom'
-import { useQuery } from 'react-query'
 
-import { authenticateUser } from '../ApiServices/AuthService'
-import { setJwt } from '../ApiServices/JwtService'
+// Authentication
+import { setJwt } from '../../ApiServices/JwtService'
+import { authenticateUser } from '../../ApiServices/AuthService'
 
-import '../styling/Login.css'
-import eyeSlash from '../assets/eye-slash.svg'
-import eye from '../assets/eye.svg'
+// Image Assets
+import eye from '../../assets/eye.svg'
+import eyeSlash from '../../assets/eye-slash.svg'
+
+// Styling
+import './LoginPage.css'
+
 
 export default function Login(){
+// initialize and states
+  const navigate = useNavigate();
   const [showPassword, setShowPassword] = useState(false);
+  const [responseMessage, setResponseMessage] = useState({state: false, error: '', msg: ''});
   const [loginForm, setLoginForm] = useState({
     email: '',
     password: ''
   });
-  const {refetch, isRefetching } = useQuery(['authentication', loginForm], ()=> authenticateUser(loginForm), {
-    enabled: false,
-    refetchOnWindowFocus: false
-  });
-  const navigate = useNavigate();
 
+
+// stores backend call that checks user authentication in refetch
+  const {refetch, isRefetching } = useQuery(
+    ['authentication', loginForm], 
+    ()=> authenticateUser(loginForm), 
+    {
+      enabled: false,
+      refetchOnWindowFocus: false,
+    }
+  );
+  
+
+// handles Input changes
   const handleFormChange = (event) => {
     const { name, value } = event.target;
-
-    setLoginForm(prevForm => {
-      return { ...prevForm, [name] : value }
-    });
+    setLoginForm(prevForm => ({ ...prevForm, [name] : value }));
   }
 
-  const handleLoginClick = async () => {
-    const loginResponse  = await refetch();
 
-    if (!loginResponse.error && loginResponse.data.jwt) {
+// calls authentication refetch on Login Click
+  const handleLoginClick = async () => {
+    // checks for missing inputs
+    if(Object.values(loginForm).indexOf('') > -1){
+      setResponseMessage({state: true, error:'missing', msg: 'Missing Input Fields'})
+      return;
+    }
+    // refetch call
+    const loginResponse  = await refetch();
+    if (!loginResponse.isError && loginResponse.data.jwt) {
       setJwt(loginResponse.data.jwt);
-      navigate('/');
+      navigate('/task-manager');
+    } else{
+      setResponseMessage({state: true, error: 'error', msg: 'Sorry, We could not find a user with this login information'})
     }
   }
 
+// show password button styling
   function password(event){setShowPassword(prevPass => !prevPass);}
   const styles = { backgroundImage: showPassword ? `url(${eye})` : `url(${eyeSlash})`}
   const types = showPassword ? 'text' : 'password';
@@ -48,18 +72,36 @@ export default function Login(){
         <Link to="/"><img className="logo" src = "https://upload.wikimedia.org/wikipedia/commons/8/84/Alternate_Task_Manager_icon.png" ></img></Link>
         <div className="sign-up-section">
           <p>Don't have an account?</p>
-          <Link to="/SignUp"><button>Sign up</button></Link>
+          <Link to="/"><button>Sign up</button></Link>
         </div>
       </header>
       <div className="input-field-container">
         <div className="input-field">
           <h1>Login</h1>
+          {
+            responseMessage.state && 
+              <div className="login-response-message">
+                <p>{responseMessage.msg}</p>
+              </div>
+          }
           <label htmlFor="email">Email</label>
-          <input onChange={(event) => handleFormChange(event)} id="email" name="email" type="email" placeholder="Enter your email" />
+          <input 
+            onChange={(event) => handleFormChange(event)} 
+            id={(responseMessage.error === 'missing' && loginForm.email === '') ? 'error' : 'email'} 
+            name="email" 
+            type="email" 
+            placeholder="Enter your email" 
+          />
 
           <label htmlFor="password">Password</label>
           <div className="password-input">
-            <input onChange={(event) => handleFormChange(event)} id="password" name="password" type={types} placeholder="Enter Password" />
+            <input 
+              onChange={(event) => handleFormChange(event)} 
+              id={(responseMessage.error === 'missing' && loginForm.password === '') ? 'error' : 'password'}
+              name="password" 
+              type={types} 
+              placeholder="Enter Password" 
+            />
             <button style={styles} onClick={() => password(event)}></button>
           </div>
